+++ workspace-service/src/graphql/folders.rs
@@ -156,7 +156,7 @@ async fn create_folder(
 #[cfg(test)]
 mod test {
     use super::*;
-    use crate::graphql::test_helpers::*;
+    use crate::graphql::test_mocks::*;
     use fnhs_event_models::EventData;
 
     #[async_std::test]
