+++ workspace-service/src/graphql/mod.rs
@@ -1,7 +1,7 @@
 mod files;
 mod folders;
 #[cfg(test)]
-mod test_helpers;
+mod test_mocks;
 mod workspaces;
 
 use super::db;
